Fix TypeScript 2.9 build
In TypeScript 2.9 the constraint of `keyof T` is now `string | number | symbol`. This is not compatible with the type of the `event` parameter in `EventEmitter` methods, which is `string | symbol`.

The compiler option `--keyofStringsOnly` reverts to TypeScript's old `keyof T` contraint, and has been added to `tsconfig.json` to resolve this breaking change.

More info: https://github.com/Microsoft/TypeScript/issues/23724