Remove --keyofStringsOnly compiler option
In the previous commit we uses this compiler option to work around a breaking change in TypeScript 2.9. This approach did not play well with other TypeScript projects that use eventemitter-ts, as it would compile the .d.ts file in this module without using --keyofStringsOnly, causing the compilation to fail.

We now use `Extract<>` to appropriately address the breaking change, as suggested in this blog post: https://www.typescriptlang.org/docs/handbook/release-notes/typescript-2-9.html